Fixed 32 vs. 64 bit issue.

+++ libosmscout-import/src/osmscout/import/GenWayDat.cpp
@@ -106,7 +106,7 @@ namespace osmscout {
       return false;
     }
 
-    writer.Write(restrictionsSet.size());
+    writer.Write((uint32_t)restrictionsSet.size());
 
     for (std::set<TurnRestrictionRef>::const_iterator r=restrictionsSet.begin();
         r!=restrictionsSet.end();
